// packages/shared-components/Select/types.ts
export type OptionShape = {
    /**
     * Текстовое представление пункта
     */
    key: string;

    /**
     * Контент, который будет отрисован в выпадающем списке и в поле при выборе
     */
    content?: ReactNode;

    /**
     * Блокирует данный пункт для выбора
     */
    disabled?: boolean;

    /**
     * Разрешает показ компонента Checkmark, иногда нужно его убирать для показа контента ошибки или пустого состояния
     */
    showCheckMark?: boolean;

    /**
     * Дополнительные данные
     */
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    value?: any;

    // Опция-прелоадер
    isPreloader?: boolean;
};
// ...
export type FieldProps = {
    id?: string;
    rightAddons?: FormControlProps['rightAddons'];
    rightAddonsCSS?: FormControlProps['rightAddonsCSS'];

    /**
     * Дополнительный класс
     */
    className?: string;

    /**
     * Выбранный пункт
     */
    selected?: OptionShape;

    /**
     * Список выбранных пунктов
     */
    selectedMultiple?: OptionShape[];

    /**
     * Метод для ручной установки выбранных пунктов
     */
    setSelectedItems: (selected: OptionShape[]) => void;

    /**
     * Метод переключающий видимость выпадающего списка
     */
    toggleMenu: () => void;

    /**
     * Флаг, можно ли выбрать несколько значений
     */
    multiple?: boolean;

    /**
     * Флаг, что есть пользовательский ввод
     */
    autocomplete?: boolean;

    /**
     * Флаг, открыто ли меню
     */
    open?: boolean;

    /**
     * Флаг, поле заблокировано
     */
    disabled?: boolean;

    /**
     * Лейбл поля
     */
    label?: ReactNode;

    wrap?: boolean;
    labelView?: 'inner' | 'outer';

    /**
     * Пропы лейбла из downshift
     */
    labelProps?: HTMLProps<HTMLLabelElement>;

    /**
     * Плейсхолдер поля
     */
    placeholder?: string;

    //   TODO: form stuff
    /** Formik meta object (inner) */
    //   meta?: FieldMetaProps<any>;

    /**
     * Подсказка под полем
     */
    hint?: ReactNode;

    error?: string;

    /**
     * Компонент стрелки
     */
    Arrow?: ReactElement | false | null;

    /**
     * Кастомный рендер выбранного пункта
     */
    valueRenderer?: BaseSelectProps['valueRenderer'];

    /**
     * Внутренние свойства, которые должны быть установлены компоненту.
     */
    innerProps: {
        onBlur?: (event: FocusEvent<HTMLDivElement | HTMLInputElement>) => void;
        onFocus?: (event: FocusEvent<HTMLDivElement | HTMLInputElement>) => void;
        onClick?: (event: MouseEvent<HTMLDivElement | HTMLInputElement>) => void;
        tabIndex: number;
        id: string;
    } & RefAttributes<HTMLDivElement | HTMLInputElement> &
        AriaAttributes;
};